The first push...
This is the work I have done so far to my hobby project Blaze. Blaze is a game launcher that uses Steam to display servers and connect directly to them.

Even though it is still under development it currently functions as a proof of concept and allows you to select a game, a server and then run the game and join that server without the need of filtering servers in the Steam server browser.

Blaze also has Discord integration to allow people to join directly through Discord and the press of a button.