1.0.1.5
Upgrades people upgrades!

The app now (sorta) installs itself! This seems to work but it still is in early testing and so could still have problems to be solved.
Limited custom games to 1 to prevent a bug (when more than 1 game is added the app hangs), started working on the removing custom game functions.